+++ plugin_ide.core/src/com/github/rustdt/ide/core/text/RustPartitionScanner.java
@@ -12,28 +12,27 @@ package com.github.rustdt.ide.core.text;
 
 import melnorme.lang.ide.core.TextSettings_Actual.LangPartitionTypes;
 import melnorme.lang.ide.core.text.PatternRule_Fixed;
+import melnorme.lang.ide.core.text.RuleBasedPartitionScannerExt;
 import melnorme.utilbox.collections.ArrayList2;
 
 import org.eclipse.jface.text.rules.IPredicateRule;
 import org.eclipse.jface.text.rules.IToken;
-import org.eclipse.jface.text.rules.RuleBasedPartitionScanner;
 import org.eclipse.jface.text.rules.Token;
 
-public class RustPartitionScanner extends RuleBasedPartitionScanner {
-	
-	private static final char NO_ESCAPE_CHAR = (char) -1;
+public class RustPartitionScanner extends RuleBasedPartitionScannerExt {
 	
 	public RustPartitionScanner() {
-		
-		ArrayList2<IPredicateRule> rules = new ArrayList2<>();
-		
-		IToken tkDocComment = new Token(LangPartitionTypes.DOC_COMMENT.getId());
-		rules.add(new PatternRule_Fixed("///", null, tkDocComment, NO_ESCAPE_CHAR, true, true));
-		rules.add(new PatternRule_Fixed("/**", "*/", tkDocComment, NO_ESCAPE_CHAR, false, true));
-		
-		IToken tkComment = new Token(LangPartitionTypes.COMMENT.getId());
-		rules.add(new PatternRule_Fixed("//", null, tkComment, NO_ESCAPE_CHAR, true, true));
-		rules.add(new PatternRule_Fixed("/*", "*/", tkComment, NO_ESCAPE_CHAR, false, true));
+	}
+	
+	@Override
+	protected void addRules(ArrayList2<IPredicateRule> rules) {
+		addStandardRules(rules, 
+			LangPartitionTypes.LINE_COMMENT.getId(), 
+			LangPartitionTypes.BLOCK_COMMENT.getId(),
+			LangPartitionTypes.DOC_LINE_COMMENT.getId(),
+			LangPartitionTypes.DOC_BLOCK_COMMENT.getId(),
+			null
+		);
 		
 		IToken tkAttribute = new Token(LangPartitionTypes.ATTRIBUTE.getId());
 		rules.add(new PatternRule_Fixed("#[", "]", tkAttribute, NO_ESCAPE_CHAR, false, true));
@@ -51,7 +50,6 @@ public class RustPartitionScanner extends RuleBasedPartitionScanner {
 		rules.add(new PatternRule_Fixed("r##\"", "\"##", tkRawString, NO_ESCAPE_CHAR, false, true));
 		rules.add(new PatternRule_Fixed("br##\"", "\"##", tkRawString, NO_ESCAPE_CHAR, false, true));
 		
-		setPredicateRules(rules.toArray(IPredicateRule.class));
 	}
 	
 }